Started refactoring the renderers.
