+++ src/Renderer/Application/Json.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * @package     Redcore
+ * @subpackage  Document
+ *
+ * @copyright   Copyright (C) 2005 - 2015 Open Source Matters, Inc. All rights reserved.
+ * @license     GNU General Public License version 2 or later; see LICENSE.txt
+ */
+
+namespace Joomla\Webservices\Renderer\Application;
+
+/**
+ * Pure JSON renderer.  Just a proxy for the HAL+JSON renderer at the moment.
+ *
+ * @package     Redcore
+ * @subpackage  Document
+ * @since       __DEPLOY_VERSION__
+ */
+class Json extends Haljson
+{
+}
